stand alone checker
