+++ watertap/core/plugins/solvers.py
@@ -13,19 +13,10 @@
 import logging
 
 import pyomo.environ as pyo
-from pyomo.common.collections import Bunch
-from pyomo.common.collections import ComponentMap, ComponentSet
-from pyomo.common.modeling import unique_component_name
 from pyomo.core.base.block import _BlockData
 from pyomo.core.kernel.block import IBlock
 from pyomo.solvers.plugins.solvers.IPOPT import IPOPT
 
-from pyomo.core.plugins.transform.add_slack_vars import AddSlackVariables
-from pyomo.core.plugins.transform.hierarchy import IsomorphicTransformation
-
-from pyomo.contrib.incidence_analysis import get_incident_variables, IncidenceMethod
-
-from pyomo.opt import WriterFactory
 
 import idaes.core.util.scaling as iscale
 from idaes.core.util.scaling import (
@@ -36,7 +27,6 @@ from idaes.core.util.scaling import (
 from idaes.logger import getLogger
 
 _log = getLogger("watertap.core")
-_default_nl_writer = WriterFactory.get_class("nl")
 
 _pyomo_nl_writer_log = logging.getLogger("pyomo.repn.plugins.nl_writer")
 
@@ -48,6 +38,10 @@ def _pyomo_nl_writer_logger_filter(record):
     return True
 
 
+@pyo.SolverFactory.register(
+    "ipopt-watertap",
+    doc="The Ipopt NLP solver, with user-based variable and automatic Jacobian constraint scaling",
+)
 class IpoptWaterTAP(IPOPT):
     def __init__(self, **kwds):
         kwds["name"] = "ipopt-watertap"
@@ -207,446 +201,6 @@ class IpoptWaterTAP(IPOPT):
                 )
             return False
         return True
-
-
-class _VariableBoundsAsConstraints(IsomorphicTransformation):
-    """Replace all variables bounds and domain information with constraints.
-
-       Leaves fixed Vars untouched (for now)
-    """
-
-    def _apply_to(self, instance, **kwds):
-
-        boundconstrblockname = unique_component_name(instance, "_variable_bounds")
-        instance.add_component(boundconstrblockname, pyo.Block())
-        boundconstrblock = instance.component(boundconstrblockname)
-
-        for v in instance.component_data_objects(pyo.Var, descend_into=True):
-            if v.fixed:
-                continue
-            lb, ub = v.bounds
-            if lb is None and ub is None:
-                continue
-            var_name = v.getname(fully_qualified=True)
-            if lb is not None:
-                con_name = "lb_for_"+var_name
-                con = pyo.Constraint(expr=(lb, v, None))
-                boundconstrblock.add_component(con_name, con)
-            if ub is not None:
-                con_name = "ub_for_"+var_name
-                con = pyo.Constraint(expr=(None, v, ub))
-                boundconstrblock.add_component(con_name, con)
-
-            # now we deactivate the variable bounds / domain
-            v.domain = pyo.Reals
-            v.setlb(None)
-            v.setub(None)
-
-
-@pyo.SolverFactory.register(
-    "ipopt-watertap",
-    doc="The Ipopt NLP solver, with user-based variable and automatic Jacobian constraint scaling",
-)
-class WaterTAPSolver:
-
-    def __init__(self, **kwds):
-
-        self.options = Bunch()
-        if kwds.get("options") is not None:
-            for key in kwds["options"]:
-                setattr(self.options, key, kwds["options"][key])
-
-    def solve(self, model, *args, **kwds):
-
-        base_solver = IpoptWaterTAP()
-
-        if "constr_viol_tol" not in self.options:
-            self.options["constr_viol_tol"] = 1e-08
-
-        for k,v in self.options.items():
-            base_solver.options[k] = v
-
-        # first, cache the values we get
-        _value_cache = ComponentMap()
-        for v in model.component_data_objects(pyo.Var, descend_into=True):
-            _value_cache[v] = v.value
-
-        try:
-            results = base_solver.solve(model, *args, **kwds)
-            if pyo.check_optimal_termination(results):
-                return results
-        except:
-            pass
-
-
-        # Something went wrong. We'll try to find a feasible point
-        # *or* prove that the model is infeasible
-        if model.parent_block() is None:
-            common_name = ""
-        else:
-            common_name = model.name
-
-        modified_model = model.clone()
-
-        _modified_model_var_to_original_model_var = ComponentMap()
-        _modified_model_value_cache = ComponentMap()
-
-        for v in model.component_data_objects(pyo.Var, descend_into=True):
-            modified_model_var = modified_model.find_component(v.name[len(common_name)+1:])
-
-            _modified_model_var_to_original_model_var[modified_model_var] = v
-            _modified_model_value_cache[modified_model_var] = _value_cache[v]
-            modified_model_var.set_value(_value_cache[v], skip_validation=True)
-
-        # TODO: For WT / IDAES models, we should probably be more
-        #       selective in *what* we elasticize. E.g., it probably
-        #       does not make sense to elasticize property calculations
-        #       and maybe certain other equality constraints calculating
-        #       values. Maybe we shouldn't elasticize *any* equality 
-        #       constraints.
-        #       For example, elasticizing the calculation of mass fraction
-        #       makes absolutely no sense and will just be noise for the
-        #       modeler to sift through. We could try to sort the constraints
-        #       such that we look for those with linear coefficients `1` on
-        #       some term and leave those be.
-        # move the variable bounds to the constraints
-        _VariableBoundsAsConstraints().apply_to(modified_model)
-
-        _var_to_constraints_map = ComponentMap()
-        _constraint_to_vars_map = ComponentMap()
-        for c in modified_model.component_data_objects(pyo.Constraint, descend_into=True):
-            _constraint_to_vars_map[c] = []
-            for v in get_incident_variables(c.body, method=IncidenceMethod.standard_repn, linear_only=False, include_fixed=False):
-                _constraint_to_vars_map[c].append(v)
-                if v not in _var_to_constraints_map:
-                    _var_to_constraints_map[v] = []
-                _var_to_constraints_map[v].append(c)
-
-        AddSlackVariables().apply_to(modified_model)
-        slack_block = modified_model._core_add_slack_variables
-
-        for v in slack_block.component_data_objects(pyo.Var):
-            v.fix(0)
-        # start with variable bounds -- these are the easist to interpret
-        for c in modified_model._variable_bounds.component_data_objects(pyo.Constraint, descend_into=True):
-            plus = slack_block.component(f"_slack_plus_{c.name}")
-            minus = slack_block.component(f"_slack_minus_{c.name}")
-            assert not (plus is None and minus is None)
-            if plus is not None:
-                plus.unfix()
-            if minus is not None:
-                minus.unfix()
-
-        # collect the initial set of infeasibilities
-        # we keep the constraint_viol_tol tight
-
-        # Phase 1 -- build the initial set of constraints, or prove feasibility
-        msg = "" 
-        fixed_slacks = ComponentSet()
-        elastic_filter = ComponentSet()
-
-        def _constraint_loop(relaxed_things, msg):
-            if msg == "":
-                msg += f"Model {model.name} may be infeasible. A feasible solution was found with only the following {relaxed_things} relaxed:\n"
-            else:
-                msg += f"Another feasible solution was found with only the following {relaxed_things} relaxed:\n"
-            while True:
-                def _constraint_generator():
-                    for v in slack_block.component_data_objects(pyo.Var):
-                        if v.value > self.options["constr_viol_tol"]:
-                            constr = _get_constraint(modified_model, v)
-                            yield constr, v.value
-                            v.fix(0)
-                            fixed_slacks.add(v)
-                            elastic_filter.add(constr)
-                msg = _get_results_with_value(_constraint_generator(), msg)
-                for var, val in _modified_model_value_cache.items():
-                    var.set_value(val, skip_validation=True)
-                results = base_solver.solve(modified_model, *args, **kwds)
-                if pyo.check_optimal_termination(results):
-                    msg += f"Another feasible solution was found with only the following {relaxed_things} relaxed:\n"
-                else:
-                    break
-            return msg
-
-
-        results = base_solver.solve(modified_model, *args, **kwds)
-        if pyo.check_optimal_termination(results):
-            msg = _constraint_loop("variable bounds", msg)
-
-        # next, try relaxing the inequality constraints
-        for v in slack_block.component_data_objects(pyo.Var):
-            c = _get_constraint(modified_model, v)
-            if c.equality:
-                # equality constraint
-                continue
-            if v not in fixed_slacks:
-                v.unfix()
-
-        results = base_solver.solve(modified_model, *args, **kwds)
-        if pyo.check_optimal_termination(results):
-            msg = _constraint_loop("inequality constraints and/or variable bounds", msg)
-
-        for v in slack_block.component_data_objects(pyo.Var):
-            if v not in fixed_slacks:
-                v.unfix()
-
-        results = base_solver.solve(modified_model, *args, **kwds)
-        # this solve should be feasible, by definition
-        if not pyo.check_optimal_termination(results):
-            msg += "Found model {model.name} to be numerically unstable."
-            raise Exception(msg)
-
-        msg = _constraint_loop("inequality constraints, equality constraints, and/or variable bounds", msg)
-
-        if len(elastic_filter) == 0:
-            # load the feasible solution into the original model
-            for modified_model_var, v in _modified_model_var_to_original_model_var.items():
-                v.set_value(modified_model_var.value, skip_validation=True)
-            base_solver.options["bound_push"] = 0.0
-            results = base_solver.solve(model, *args, **kwds)
-            return results
-
-        # Phase 2 -- deletion filter
-        # temporarily switch to nl_v1 writer
-        WriterFactory.register("nl")(WriterFactory.get_class("nl_v1"))
-
-        # remove slacks by fixing them to 0
-        for v in slack_block.component_data_objects(pyo.Var):
-            v.fix(0)
-        for o in modified_model.component_data_objects(pyo.Objective, descend_into=True):
-            o.deactivate()
-
-        # mark all constraints not in the filter as inactive
-        for c in modified_model.component_data_objects(pyo.Constraint):
-            if c in elastic_filter:
-                continue
-            else:
-                c.deactivate()
-
-        try:
-            results = base_solver.solve(modified_model, *args, **kwds)
-        except:
-            results = None
-        
-        if pyo.check_optimal_termination(results):
-            msg += "Could not determine Minimal Intractable System\n"
-        else:
-            deletion_filter = []
-            guards = []
-            for constr in elastic_filter:
-                constr.deactivate()
-                for var, val in _modified_model_value_cache.items():
-                    var.set_value(val, skip_validation=True)
-                try:
-                    results = base_solver.solve(modified_model, *args, **kwds)
-                except:
-                    math_failure = True
-                else:
-                    math_failure = False
-
-                if math_failure:
-                    constr.activate()
-                    guards.append(constr)
-                elif pyo.check_optimal_termination(results):
-                    constr.activate()
-                    deletion_filter.append(constr)
-                else: # still infeasible without this constraint
-                    pass
-
-            msg += "Computed Minimal Intractable System (MIS)!\n"
-            msg += "Constraints / bounds in MIS:\n"
-            msg = _get_results(deletion_filter, msg)
-            msg += "Constraints / bounds in guards for stability:"
-            msg = _get_results(guards, msg)
-
-        WriterFactory.register("nl")(_default_nl_writer)
-
-        raise Exception(msg)
-
-        # TODO: Elasticizing too much at once seems to cause Ipopt trouble.
-        #       After an initial sweep, we should just fix one elastic variable
-        #       and put everything else on a stack of "constraints to elasticize".
-        #       We elastisize one constraint at a time and fix one constraint at a time.
-        #       After fixing an elastic variable, we elasticize a single constraint it
-        #       appears in and put the remaining constraints on the stack. If the resulting problem
-        #       is feasible, we keep going "down the tree". If the resulting problem is
-        #       infeasible or cannot be solved, we elasticize a single constraint from
-        #       the top of the stack.
-        #       The algorithm stops when the stack is empty and the subproblem is infeasible.
-        #       Along the way, any time the current problem is infeasible we can check to
-        #       see if the current set of constraints in the filter is as a collection of
-        #       infeasible constraints -- to terminate early.
-        stack = []
-        elastic_filter = ComponentSet()
-        proved_feasibility = False
-        while pyo.check_optimal_termination(results) or stack:
-            if pyo.check_optimal_termination(results):
-                max_viol = 0.0
-                max_pair = None
-                for v in slack_block.component_data_objects(pyo.Var):
-                    if v.value > self.options["constr_viol_tol"]:
-                        constr = _get_constraint(modified_model, v)
-                        stack.append((v, constr))
-                        if v.value > max_viol:
-                            max_viol = v.value
-                            max_pair = (v, constr)
-            else: # infeasible, pop
-                max_pair = stack.pop()
-            if max_pair is None:
-                proved_feasibility = True
-                break
-            else:
-                max_pair[0].fix(0)
-                elastic_filter.add(max_pair[1])
-                print(f"Adding constraint {max_pair[1].name} to the filter")
-            # for var, val in _modified_model_value_cache.items():
-            #     var.set_value(val, skip_validation=True)
-            # fix everything but the max, then only unfix those
-            # in the graph related to the enforced constraint
-            for v in slack_block.component_data_objects(pyo.Var):
-                if v.value > self.options["constr_viol_tol"]:
-                    continue
-                else:
-                    for vs, _ in stack:
-                        if vs is v:
-                            break
-                    else: # no break
-                        v.fix(0)
-            for c in [max_pair[1]]:
-                for v in _constraint_to_vars_map[c]:
-                    for c_prime in _var_to_constraints_map[v]:
-                        if c_prime not in elastic_filter:
-                            plus = slack_block.component(f"_slack_plus_{c_prime.name}")
-                            minus = slack_block.component(f"_slack_minus_{c_prime.name}")
-                            assert not (plus is None and minus is None)
-                            if plus is not None:
-                                plus.unfix()
-                            if minus is not None:
-                                minus.unfix()
-                            print(f"relaxing constraint {c_prime.name}")
-            try:
-                results = base_solver.solve(modified_model, *args, **kwds)
-            except:
-                break
-
-        if proved_feasibility:
-            # load the feasible solution into the original model
-            for modified_model_var, v in _modified_model_var_to_original_model_var.items():
-                v.set_value(modified_model_var.value, skip_validation=True)
-            base_solver.options["bound_push"] = 0.0
-            results = base_solver.solve(model, *args, **kwds)
-            return results
-
-        # Phase 2 -- deletion filter
-        print(f"PHASE 2, {len(elastic_filter)} Constraints to consider.")
-        # temporarily switch to nl_v1 writer
-        WriterFactory.register("nl")(WriterFactory.get_class("nl_v1"))
-
-        # remove slacks by fixing them to 0
-        for v in slack_block.component_data_objects(pyo.Var):
-            v.fix(0)
-        for o in modified_model.component_data_objects(pyo.Objective, descend_into=True):
-            o.deactivate()
-
-        # mark all constraints not in the filter as inactive
-        for c in modified_model.component_data_objects(pyo.Constraint):
-            if c in elastic_filter:
-                continue
-            else:
-                c.deactivate()
-
-        # TODO: From here on, change Ipopt's options
-        #       so it will terminate on a feasible solution
-        #       (e.g., do not consider dual or complimentary
-        #              or overall `tol`)
-        base_solver.options["tol"] = 1e-08
-        base_solver.options["dual_inf_tol"] = 1e20
-        base_solver.options["compl_inf_tol"] = 1e20
-        try:
-            results = base_solver.solve(modified_model, *args, **kwds)
-        except:
-            results = None
-        assert results is None or not pyo.check_optimal_termination(results)
-
-        deletion_filter = []
-        guards = []
-        for constr in elastic_filter:
-            constr.deactivate()
-            for var, val in _modified_model_value_cache.items():
-                var.set_value(val, skip_validation=True)
-
-            try:
-                results = base_solver.solve(modified_model, *args, **kwds)
-            except:
-                math_failure = True
-            else:
-                math_failure = False
-
-            if math_failure:
-                constr.activate()
-                guards.append(constr)
-            elif pyo.check_optimal_termination(results):
-                constr.activate()
-                deletion_filter.append(constr)
-            else: # still infeasible without this constraint
-                pass
-
-        print("Constraints / bounds in set:")
-        print(_get_results(deletion_filter))
-        print("Constraints / bounds in guards:")
-        print(_get_results(guards))
-        WriterFactory.register("nl")(_default_nl_writer)
-
-        raise Exception("Found model to be infeasible, see conflicting constraints above")
-
-def _get_results_with_value(constr_value_generator, msg=None):
-    if msg is None:
-        msg = ""
-    for c, value in constr_value_generator:
-        c_name = c.name
-        if "_variable_bounds" in c_name:
-            name = c.local_name
-            if "lb" in name:
-                msg += f"\tlb of var {name[7:]} by {value}\n"
-            elif "ub" in name:
-                msg += f"\tub of var {name[7:]} by {value}\n"
-            else:
-                raise RuntimeError("unrecongized var name")
-        else:
-            msg += f"\tconstraint: {c_name} by {value}\n"
-    return msg
-
-def _get_results(constr_generator, msg=None):
-    if msg is None:
-        msg = ""
-    for c in constr_generator:
-        c_name = c.name
-        if "_variable_bounds" in c_name:
-            name = c.local_name
-            if "lb" in name:
-                msg += f"\tlb of var {name[7:]}\n"
-            elif "ub" in name:
-                msg += f"\tub of var {name[7:]}\n"
-            else:
-                raise RuntimeError("unrecongized var name")
-        else:
-            msg += f"\tconstraint: {c_name}\n"
-    return msg
-
-def _get_constraint(modified_model, v):
-    if "_slack_plus_" in v.name:
-        constr = modified_model.find_component(v.local_name[len("_slack_plus_"):])
-        if constr is None:
-            raise RuntimeError("Bad constraint name {v.local_name[len('_slack_plus_'):]}")
-        return constr
-    elif "_slack_minus_" in v.name:
-        constr = modified_model.find_component(v.local_name[len("_slack_minus_"):])
-        if constr is None:
-            raise RuntimeError("Bad constraint name {v.local_name[len('_slack_minus_'):]}")
-        return constr
-    else:
-        raise RuntimeError("Bad var name {v.name}")
 
 
 ## reconfigure IDAES to use the ipopt-watertap solver
